+++ buildVars.py
@@ -40,7 +40,7 @@ addon_info = AddonInfo(
 	addon_sourceURL= "https://github.com/Accessolutions/telenvda-accessolutions",
 	# Brief changelog for this version
 	# Translators: what's new content for the add-on version to be shown in the add-on store
-	addon_changelog=_("""Add WebSocket relay connections over HTTPS, including port 443, add Windows SSPI authentication for NTLM and Kerberos HTTP proxies, improve proxy compatibility, and provide two remote screenshot methods."""),
+	addon_changelog=_("""Fix automatic updates that failed to complete with an "access denied" error, recover installations stuck pending, add WebSocket relay connections over HTTPS, including port 443, add Windows SSPI authentication for NTLM and Kerberos HTTP proxies, improve proxy compatibility, and provide two remote screenshot methods."""),
 )
 
 import os.path